Move sitemesh specific tag libraries that require the sitemesh2 dependency into a separate sitmesh2 specific plugin that can optionally not be used.
