+++ gradle.properties
@@ -1,6 +1,6 @@
 title=Groovy Server Pages (GSP)
 authors=Puneet Behl
-projectVersion=6.0.1-SNAPSHOT
+projectVersion=6.0.3-SNAPSHOT
 projectDesc=GSP (Groovy Server Pages) - A server-side view rendering technology based on Groovy
 projectUrl=https://github.com/grails/grails-gsp
 githubSlug=grails/grails-gsp
