+++ grails-plugin-gsp/src/test/groovy/org/grails/web/taglib/AbstractGrailsTagTests.groovy
@@ -270,6 +270,7 @@ info.app.name: ${getClass().name}
         dependantPluginClasses << gcl.loadClass("org.grails.plugins.i18n.I18nGrailsPlugin")
         dependantPluginClasses << gcl.loadClass("org.grails.plugins.web.mapping.UrlMappingsGrailsPlugin")
         dependantPluginClasses << gcl.loadClass("org.grails.plugins.web.controllers.ControllersGrailsPlugin")
+        dependantPluginClasses << gcl.loadClass("org.grails.plugins.web.Sitemesh2GrailsPlugin")
         dependantPluginClasses << gcl.loadClass("org.grails.plugins.web.GroovyPagesGrailsPlugin")
 
         def dependentPlugins = dependantPluginClasses.collect { new DefaultGrailsPlugin(it, grailsApplication)}
