+++ settings.gradle
@@ -32,6 +32,7 @@ rootProject.name = "grails-gsp"
 
 include 'grails-gsp'
 include 'grails-plugin-gsp'
+include 'grails-plugin-sitemesh2'
 include 'grails-taglib'
 include 'grails-web-gsp'
 include 'grails-web-gsp-taglib'
